delete duplicate code table core unit
