+++ src/stories/containers/cu-table/cu-table.mvvm.ts
@@ -319,5 +319,6 @@ export const useCoreUnitsTableMvvm = () => {
     onSortClick,
     applySort,
     queryStrings,
+    sortData,
   };
 };
